refactor: structure of project refactored based on standards fixed #128
