+++ cmd/cli/main.go
@@ -5,13 +5,14 @@ import (
 	"os"
 	"strings"
 
+	"github.com/pagu-project/Pagu/internal/engine"
+	"github.com/pagu-project/Pagu/internal/engine/command"
+	"github.com/pagu-project/Pagu/pkg/log"
+
 	"github.com/pactus-project/pactus/crypto"
 	pagu "github.com/pagu-project/Pagu"
 	pCmd "github.com/pagu-project/Pagu/cmd"
 	"github.com/pagu-project/Pagu/config"
-	"github.com/pagu-project/Pagu/engine"
-	"github.com/pagu-project/Pagu/engine/command"
-	"github.com/pagu-project/Pagu/log"
 	cobra "github.com/spf13/cobra"
 )
