+++ internal/delivery/grpc/buf/buf.gen.yaml
