+++ internal/engine/interface.go
@@ -1,0 +1,10 @@
+package engine
+
+import (
+	"github.com/pagu-project/Pagu/internal/engine/command"
+)
+
+type IEngine interface {
+	Run(appID command.AppID, callerID string, tokens []string) (*command.CommandResult, error)
+	Commands() []command.Command
+}
